Create Netflix Data Analysis
In this project, I combined all the skill I have learned thus far when dealing with relational data. I was able to query the databases for information needed in the most efficient and correct way. This is also the first time I used PostgreSQL so there was a small bit of syntax change that I was able to quickly adapt to, coming from Standard SQL.